2-sol/requirements.txt
alabaster==0.7.13
asgiref==3.6.0
autopep8==2.0.2
Babel==2.12.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.52.0
django-environ==0.10.0
django-extensions==3.2.1
django-jet==1.0.8
django-jet-4==0.0.1
django-rest-swagger==2.2.0
django-storages==1.13.2
django_debug_toolbar==3.8.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
docutils==0.19
drf-yasg==1.20.0
dropbox==11.36.0
future==0.18.3
gunicorn==20.1.0
idna==3.4
imagesize==1.4.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.0
pdoc3==0.10.0
Pillow==9.4.0
ply==3.11
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
serializers==0.2.4
simplejson==3.18.3
six==1.16.0
snowballstemmer==2.2.0
Sphinx==6.1.3
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.3
stone==3.3.1
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.15
uWSGI==2.0.21
whitenoise==6.4.0